requirements.txt
fastapi
uvicorn
scikit-learn
pandas
mlflow
joblib
pyyaml
evidently